+++ CarBudget.cpp
@@ -25,13 +25,13 @@
 
 #include <sailfishapp.h>
 #include <QSettings>
-#include "tank.h"
-#include "cost.h"
-#include "tire.h"
-#include "station.h"
-#include "car.h"
-#include "carmanager.h"
-#include "filemodel.h"
+#include "src/tank.h"
+#include "src/cost.h"
+#include "src/tire.h"
+#include "src/station.h"
+#include "src/car.h"
+#include "src/carmanager.h"
+#include "src/filemodel.h"
 #include "qmlLibs/qquickfolderlistmodel.h"
 
 #include <QtCore/QTranslator>
