+++ src/carevent.cpp
@@ -1,5 +1,5 @@
 #include "carevent.h"
-#include <car.h>
+#include "car.h"
 
 CarEvent::CarEvent(Car *parent):
     QObject(parent),
