+++ src/globals.cpp
