+++ src/statfileinfo.cpp
