+++ src/statfileinfo.h
